+++ config/routes.php
@@ -10,6 +10,7 @@
  * @link        https://github.com/mirko-pagliai/cakephp-assets
  * @license     https://opensource.org/licenses/mit-license.php MIT License
  */
+
 use Cake\Routing\RouteBuilder;
 use Cake\Routing\Router;