+++ src/Utility/AssetsCreator.php
@@ -17,8 +17,8 @@ use Cake\Core\Configure;
 use Cake\Core\Plugin;
 use Cake\Filesystem\File;
 use InvalidArgumentException;
-use MatthiasMullie\Minify;
-use RuntimeException;
+use MatthiasMullie\Minify\CSS;
+use MatthiasMullie\Minify\JS;
 
 /**
  * An utility to create assets
@@ -29,7 +29,7 @@ class AssetsCreator
      * Asset full path
      * @var string
      */
-    protected $asset = null;
+    protected $asset;
 
     /**
      * File paths that will be transformed into a single asset
@@ -41,14 +41,13 @@ class AssetsCreator
      * Asset type (`css` or `js`)
      * @var string
      */
-    protected $type = null;
+    protected $type;
 
     /**
      * Construct. Sets the asset type and paths
      * @param string|array $paths String or array of css files
      * @param string $type Extension (`css` or `js`)
-     * @return $this
-     * @throws InvalidArgumentException
+     * @throws \InvalidArgumentException
      * @uses resolveAssetPath()
      * @uses resolveFilePaths()
      * @uses $asset
@@ -57,17 +56,17 @@ class AssetsCreator
      */
     public function __construct($paths, $type)
     {
-        if (!in_array($type, ['css', 'js'])) {
-            throw new InvalidArgumentException(__d('assets', 'Asset type `{0}` not supported', $type));
-        }
+        is_true_or_fail(
+            in_array($type, ['css', 'js']),
+            __d('assets', 'Asset type `{0}` not supported', $type),
+            InvalidArgumentException::class
+        );
 
         //Note: `resolveFilePaths()` method needs `$type` property;
         //  `resolveAssetPath()` method needs `$type` and `$paths` properties
         $this->type = $type;
         $this->paths = $this->resolveFilePaths($paths);
         $this->asset = $this->resolveAssetPath();
-
-        return $this;
     }
 
     /**
@@ -110,7 +109,6 @@ class AssetsCreator
 
             //Appends the file extension, if not already present
             $path = pathinfo($path, PATHINFO_EXTENSION) == $this->type ? $path : sprintf('%s.%s', $path, $this->type);
-
             is_readable_or_fail($path);
 
             return $path;
@@ -120,43 +118,24 @@ class AssetsCreator
     /**
      * Creates the asset
      * @return string
-     * @throws RuntimeException
-     * @uses filename()
      * @uses path()
      * @uses $paths
      * @uses $type
      */
     public function create()
     {
-        if (!is_readable($this->path())) {
-            switch ($this->type) {
-                case 'css':
-                    $minifier = new Minify\CSS();
-                    break;
-                case 'js':
-                    $minifier = new Minify\JS();
-                    break;
-            }
+        $File = new File($this->path());
 
+        if (!$File->exists() || !$File->readable()) {
+            $minifier = $this->type === 'css' ? new CSS() : new JS();
             array_map([$minifier, 'add'], $this->paths);
 
             //Writes the file
-            if (!(new File($this->path(), false, 0755))->write($minifier->minify())) {
-                throw new RuntimeException(__d('assets', 'Failed to create file {0}', rtr($this->path())));
-            }
+            $success = $File->Folder->pwd() && $File->write($minifier->minify());
+            is_true_or_fail($success, __d('assets', 'Failed to create file {0}', rtr($this->path())));
         }
 
-        return $this->filename();
-    }
-
-    /**
-     * Returns the asset filename
-     * @return string Asset filename
-     * @uses path()
-     */
-    public function filename()
-    {
-        return pathinfo($this->path(), PATHINFO_FILENAME);
+        return $File->name();
     }
 
     /**