+++ tests/TestCase/View/Helper/AssetHelperTest.php
@@ -47,8 +47,8 @@ class AssetHelperTest extends TestCase
             'Assets.timestamp' => true,
         ]);
 
-        $this->Asset = new AssetHelper(new View);
-        $this->Html = new HtmlHelper(new View);
+        $this->Asset = new AssetHelper(new View());
+        $this->Html = new HtmlHelper(new View());
     }
 
     /**