+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # 1.x branch
 ## 1.4 branch
+### 1.4.2
+* removed useless `AssetsCreator::filename()` method;
+* added [API](//mirko-pagliai.github.io/cakephp-assets).
+
+### 1.4.1
+* updated for `php-tools` 1.1.12.
+
 ### 1.4.0
 * removed `ASSETS` contants;
 * removed `IntegrationTestCase` class, use instead `IntegrationTestTrait`